requirements.txt
langchain-core>=0.1.0  # Core LangChain components
langchain-community>=0.0.16  # Community components
langchain-google-genai>=0.0.4  # For Google Generative AI integration
google-auth>=2.23.4  # For Google OAuth credentials handling
python-dotenv>=1.0.0  # For loading environment variables
streamlit>=1.26.0